RFC: Move validate marshal and errors from utils to code
