+++ core/contractsdk/go/code/bigint.go
@@ -1,4 +1,4 @@
-package utils
+package code
 
 import "math/big"
