+++ core/contractsdk/go/code/errors.go
@@ -1,4 +1,4 @@
-package utils
+package code
 
 import "errors"
