+++ core/contractsdk/go/code/validate.go
@@ -1,4 +1,4 @@
-package utils
+package code
 
 import (
 	"fmt"
