+++ core/contractsdk/go/code/validate_test.go
@@ -1,4 +1,4 @@
-package utils
+package code
 
 import (
 	"math/big"
