+++ core/contractsdk/go/example/score_record/main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"github.com/xuperchain/xuperchain/core/contractsdk/go/code"
 	"github.com/xuperchain/xuperchain/core/contractsdk/go/driver"
-	"github.com/xuperchain/xuperchain/core/contractsdk/go/utils"
 )
 
 const (
@@ -19,7 +18,7 @@ func (sr *scoreRecord) Initialize(ctx code.Context) code.Response {
 		Owner string `json:"owner" validte:"required"`
 	}{}
 
-	if err := utils.Unmarshal(ctx.Args(), &args); err != nil {
+	if err := code.Unmarshal(ctx.Args(), &args); err != nil {
 		return code.Error(err)
 	}
 	if err := ctx.PutObject([]byte(OWNER_KEY), []byte(args.Owner)); err != nil {
@@ -31,21 +30,21 @@ func (sr *scoreRecord) Initialize(ctx code.Context) code.Response {
 func (sc *scoreRecord) AddScore(ctx code.Context) code.Response {
 	initiator := ctx.Initiator()
 	if initiator == "" {
-		return code.Error(utils.ErrMissingInitiator)
+		return code.Error(code.ErrMissingInitiator)
 	}
 	owner, err := ctx.GetObject([]byte(OWNER_KEY))
 	if err != nil {
 		return code.Error(err)
 	}
 	if string(owner) != initiator {
-		return code.Error(utils.ErrPermissionDenied)
+		return code.Error(code.ErrPermissionDenied)
 	}
 	args := struct {
 		UserId string `json:"user_id" validte:"required"`
 		Data   string `json:"data" validte:"required"`
 	}{}
 
-	if err := utils.Unmarshal(ctx.Args(), &args); err != nil {
+	if err := code.Unmarshal(ctx.Args(), &args); err != nil {
 		return code.Error(err)
 	}
 
@@ -59,7 +58,7 @@ func (sr *scoreRecord) QueryScore(ctx code.Context) code.Response {
 	args := struct {
 		UserId string `json:"user_id" validte:"required"`
 	}{}
-	if err := utils.Unmarshal(ctx.Args(), &args); err != nil {
+	if err := code.Unmarshal(ctx.Args(), &args); err != nil {
 		return code.Error(err)
 	}
 	data, err := ctx.GetObject([]byte(RECORD_KEY + args.UserId))
