+++ core/contractsdk/go/example/scripts/build.sh
@@ -1,5 +1,6 @@
 for dir in `ls`;do
   if [ -f ${dir}/main.go ] ;then
+    echo building $dir ...
     GOOS=js GOARCH=wasm go build -o wasm/${dir}.wasm ${dir}/main.go
   fi
 done
